refactor: extract data import logic
Data import takes too many methods in DatabaseController, leading to a break of the Single Responsability Principle. Therefore, this commit starts the process of extracting this logic into a new module called data_import.
